\name{CASES99}
\alias{CASES99}
\docType{data}
\title{
One day of 1-s averages of the thermocouple temperature data from CASES-99 dataset
}
\description{
These are 1-s averages of the CASES-99 (Poulos et al. 2002) thermocouple temperature data at the seventh level (9.5 m) from 1100 LST 5 October to 1100 LST 6 October.
}
\usage{data(CASES99)}

\details{
Cooperative Atmosphere-Surface Exchange Study (CASES-99) was conducted over a relatively flat-terrain rural grassland site near Leon, Kansas, during October 1999. As a part of the extensive observations, a 60-m tower was equipped with thermocouples at 34 vertical levels (0.23, 0.63, 2.3 m, and every 1.8 m above 2.3 m) that sampled air temperature five times per second (Sun et al. 2012), while 20-Hz sonic anemometer measurements were taken at seven levels (1.5, 5, 10, 20, 30, 40, 50, and 55 m). 1-s averages of the CASES-99 thermocouple temperature data at the seventh level (9.5 m) from 1100 LST 5 October to 1100 LST 6 October are taken as an example for detection and  clustering of events. 
}



\source{
Gregory S. Poulos, William Blumen, David C. Fritts, Julie
K. Lundquist, Jielun Sun, Sean P. Burns, Carmen Nappo,
Robert Banta, Rob Newsom, Joan Cuxart, Enric Terradellas,
Ben Balsley, and Michael Jensen. CASES-99: A comprehensive
investigation of the stable nocturnal boundary layer
(2002). \emph{Bulletin of the American Meteorological
Society}, \bold{83}(4):555-581.
}
\examples{
data(CASES99)
}
\keyword{datasets}
