% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A simple greeting function}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{the text to display, default \code{"world"}}
}
\description{
This function shows a standard text on the console. In a time-honoured
tradition, it defaults to displaying \emph{hello, world}.
}
\examples{
  hello()
  hello("and goodbye")
}
