% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicplot.R
\name{simple_plot}
\alias{simple_plot}
\title{A simple scatterplot using ggplot2}
\usage{
simple_plot(x, y)
}
\arguments{
\item{x}{the abscissa values}

\item{y}{the ordinate values}
}
\value{
a ggplot object
}
\description{
A simple scatterplot using ggplot2
}
\examples{
simple_plot(mtcars$wt, mtcars$mpg)
}
