\name{ur.za.fast}
\alias{ur.za.fast}
\title{Unit root test for events considering a structrual break}
\usage{
ur.za.fast(y, model = c("intercept", "trend", "both"), lag = NULL)
}
\arguments{
  \item{y}{a time series}

  \item{model}{Three choices: 'intercept', 'trend' or
  'both')}

  \item{lag}{a scalar chosen as lag}
}
\description{
This function performs unit root test for events
considering a structrual break.
}

